Enable EnforceCodeStyleInBuild
The reason to drop `--no-restore` from the lint step is that enabling
EnforceCodeStyleInBuild is affecting dotnet format so it now performs
checks that require a restore. (The lint step started to fail.)

There might be some overlap between the build and the lint step now, and
calling just `dotnet format whitespace` _might_ be enough now that the
build is enforcing code style. But I have hard time verifying if that is
the case, and I rather have some redundancy than miss enforcing code
style and code analysis checks.